generalize TwitterSourceOp
