+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/apis/twitter/TwitterSourceOpExec.scala
@@ -1,0 +1,26 @@
+package edu.uci.ics.texera.workflow.operators.source.apis.twitter
+import com.github.redouane59.twitter.TwitterClient
+import com.github.redouane59.twitter.signature.TwitterCredentials
+import edu.uci.ics.texera.workflow.common.operators.source.SourceOperatorExecutor
+
+abstract class TwitterSourceOpExec(
+    accessToken: String,
+    accessTokenSecret: String,
+    apiKey: String,
+    apiSecretKey: String
+) extends SourceOperatorExecutor {
+  var twitterClient: TwitterClient = _
+  override def open(): Unit = {
+    twitterClient = new TwitterClient(
+      TwitterCredentials
+        .builder()
+        .accessToken(accessToken)
+        .accessTokenSecret(accessTokenSecret)
+        .apiKey(apiKey)
+        .apiSecretKey(apiSecretKey)
+        .build()
+    )
+  }
+
+  override def close(): Unit = {}
+}
