+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/apis/twitter/v2/TwitterFullArchiveSearchSourceOpExecConfig.scala
@@ -1,15 +1,11 @@
-package edu.uci.ics.texera.workflow.operators.source.apis
-import edu.uci.ics.amber.engine.architecture.breakpoint.globalbreakpoint.GlobalBreakpoint
+package edu.uci.ics.texera.workflow.operators.source.apis.twitter.v2
+
 import edu.uci.ics.amber.engine.architecture.deploysemantics.deploymentfilter.UseAll
 import edu.uci.ics.amber.engine.architecture.deploysemantics.deploystrategy.RoundRobinDeployment
 import edu.uci.ics.amber.engine.architecture.deploysemantics.layer.WorkerLayer
-import edu.uci.ics.amber.engine.common.virtualidentity.{
-  ActorVirtualIdentity,
-  LayerIdentity,
-  OperatorIdentity
-}
-import edu.uci.ics.amber.engine.operators.OpExecConfig
+import edu.uci.ics.amber.engine.common.virtualidentity.{LayerIdentity, OperatorIdentity}
 import edu.uci.ics.texera.workflow.common.tuple.schema.Schema
+import edu.uci.ics.texera.workflow.operators.source.apis.twitter.TwitterSourceOpExecConfig
 
 class TwitterFullArchiveSearchSourceOpExecConfig(
     operatorIdentifier: OperatorIdentity,
@@ -20,7 +16,7 @@ class TwitterFullArchiveSearchSourceOpExecConfig(
     apiKey: String,
     apiSecretKey: String,
     searchQuery: String
-) extends OpExecConfig(operatorIdentifier) {
+) extends TwitterSourceOpExecConfig(operatorIdentifier) {
 
   override lazy val topology: Topology = {
     new Topology(
@@ -44,11 +40,5 @@ class TwitterFullArchiveSearchSourceOpExecConfig(
       ),
       Array()
     )
-  }
-
-  override def assignBreakpoint(
-      breakpoint: GlobalBreakpoint[_]
-  ): Array[ActorVirtualIdentity] = {
-    topology.layers(0).identifiers
   }
 }
